+++ CHANGES.rst
@@ -6,9 +6,21 @@ Changelog
 - new PATCH version for backwards compatible bug fixes
 
 
+0.3.0
+------
+2020-07-22: feature release
+    - autodetect travis settings
+    - autodetect binder/jupyter settings
+
+0.2.0
+------
+2020-07-22: feature release
+    - log_exception_traceback and print_exception_traceback will also report stdout, stderr if present
+
+
 0.1.4
 -------
-2020-07-17: development
+2020-07-17: feature release
     - bump coverage
 
 0.1.3